Fix looking up 'XDG_CONFIG_HOME' envvar
README says that 'era' lookups "$XDG_CONFIG_HOME" as a configuration
directory, but it used to just lookup "$HOME/.config". I guess this
doesn't make sense. So 'era' first lookups "$XDG_CONFIG_HOME", and if
not found the variable, 'era' assumes the value is ~/.config/era now.
This behavior probably is familiar in CLI application configurations.